仕様変更 id の型を int から long に変更
データベースから返ってくる型が long であるため、int ではキャストできないため

+++ TodoManager2/model/Todo.cs
@@ -12,11 +12,11 @@ namespace TodoManager2.model {
 
         }
 
-        public Todo(DateTime creationDate, int uniqueID) {
+        public Todo(DateTime creationDate, long uniqueID) {
             CreationDateTime = creationDate;
         }
 
-        public Todo(DateTime creationDate, DateTime completionDate, int uniqueID) : this(creationDate, uniqueID) {
+        public Todo(DateTime creationDate, DateTime completionDate, long uniqueID) : this(creationDate, uniqueID) {
             CompletionDateTime = completionDate;
         }
 
@@ -63,7 +63,7 @@ namespace TodoManager2.model {
         /// データベースに定義されているプライマリーキーのIDと同じ値です。
         /// 新規でオブジェクトを作成した場合のデフォルトは０になっています。
         /// </summary>
-        public int ID { get; set; } = 0;
+        public long ID { get; set; } = 0;
 
     }
 }
